# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vsy Sports API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List user bookings (or all for admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a turf slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/payment:
    post:
      operationId: createBookingPayment
      tags: [bookings]
      summary: Create payment for booking (Razorpay)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
  /bookings/{id}/verify-payment:
    post:
      operationId: verifyBookingPayment
      tags: [bookings]
      summary: Verify Razorpay payment signature
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentBody"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # COMMUNITY
...
  /admin/users/{id}/block:
    put:
      operationId: blockUser
      tags: [admin]
      summary: Block or unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}/role:
    put:
      operationId: updateUserRole
      tags: [admin]
      summary: Update user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/turfs/{id}/approve:
    put:
      operationId: approveTurf
      tags: [admin]
      summary: Approve or reject a turf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveTurfBody"
      responses:
        "200":
          description: Updated turf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turf"
  /admin/turfs/{id}/feature:
    put:
      operationId: featureTurf
      tags: [admin]
      summary: Feature or unfeature a turf
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureTurfBody"
      responses:
        "200":
          description: Updated turf
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turf"
  /admin/revenue:
    get:
      operationId: getRevenueStats
      tags: [admin]
      summary: Get revenue and booking trends (charts)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, year]
      responses:
        "200":
          description: Revenue stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueStats"
components:
  schemas:
...
    RegisterBody:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [user, turf_owner]
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, turf_owner, admin]
        phone:
          type: string
        avatar:
          type: string
        blocked:
          type: boolean
        createdAt:
          type: string

    # TURFS
    Turf:
      type: object
      required: [id, name, pricePerHour, ownerId]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        pricePerHour:
          type: number
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        address:
          type: string
        area:
          type: string
        amenities:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        featured:
          type: boolean
        ownerId:
          type: integer
        ownerName:
          type: string
        distanceKm:
          type: number
        createdAt:
          type: string
...
    Booking:
      type: object
      required: [id, turfId, userId, slotId, date, totalPrice, status]
      properties:
        id:
          type: integer
        turfId:
          type: integer
        turfName:
          type: string
        turfImage:
          type: string
        turfArea:
          type: string
        userId:
          type: integer
        userName:
          type: string
        slotId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        date:
          type: string
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        paymentStatus:
          type: string
          enum: [unpaid, paid]
        createdAt:
          type: string
    CreateBookingBody:
      type: object
      required: [turfId, slotId, date]
      properties:
        turfId:
          type: integer
        slotId:
          type: integer
        date:
          type: string
    PaymentOrder:
      type: object
      required: [orderId, amount, currency, key]
      properties:
        orderId:
          type: string
        amount:
          type: number
        currency:
          type: string
        key:
          type: string
    VerifyPaymentBody:
      type: object
      required: [razorpayOrderId, razorpayPaymentId, razorpaySignature]
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string

    # COMMUNITY
...
    RevenueStats:
      type: object
      required: [bookingRevenue, shopRevenue, chartData]
      properties:
        bookingRevenue:
          type: number
        shopRevenue:
          type: number
        chartData:
          type: array
          items:
            $ref: "#/components/schemas/ChartPoint"
    ChartPoint:
      type: object
      required: [label, bookings, revenue]
      properties:
        label:
          type: string
        bookings:
          type: integer
        revenue:
          type: number
    BlockUserBody:
      type: object
      required: [blocked]
      properties:
        blocked:
          type: boolean
    UpdateRoleBody:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [user, turf_owner, admin]
    ApproveTurfBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
    FeatureTurfBody:
      type: object
      required: [featured]
      properties:
        featured:
          type: boolean