# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vsy Sports API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPaymentStatus } from "./bookingPaymentStatus";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  turfId: number;
  turfName?: string;
  turfImage?: string;
  turfArea?: string;
  userId: number;
  userName?: string;
  slotId: number;
  startTime?: string;
  endTime?: string;
  date: string;
  totalPrice: number;
  status: BookingStatus;
  paymentStatus?: BookingPaymentStatus;
  createdAt?: string;
}